okProtectedAPI

+++ server/routes/accounts_flag/[id].delete.ts
@@ -1,7 +1,10 @@
 import { accountsFlagService } from '~/server/services/accountsFlagService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return accountsFlagService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return accountsFlagService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
