+++ server/routes/archive/[id].delete.ts
@@ -1,8 +1,10 @@
 import { archivesService } from '~/server/services/archivesService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-
-	const id = event.context.params.id
-	return archivesService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return archivesService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
