+++ server/routes/clubhouse/[id].delete.ts
@@ -1,9 +1,10 @@
 import { clubhouseService } from '~/server/services/clubhouseService'
 
 export default defineEventHandler(async (event) => {
-	const config = useRuntimeConfig()
-	const headers = event.node.req.headers
-
 	const id = event.context.params.id
-	return clubhouseService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		return clubhouseService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
