+++ server/routes/content/[id].delete.ts
@@ -2,5 +2,9 @@ import { contentService } from '~/server/services/contentService'
 
 export default defineEventHandler(async (event) => {
 	const id = event.context.params.id
-	return contentService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		return contentService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
