+++ server/routes/contributions/[id].delete.ts
@@ -1,7 +1,10 @@
 import { contributionsService } from '~/server/services/contributionsService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return contributionsService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return contributionsService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
