+++ server/routes/events/[id].delete.ts
@@ -1,8 +1,10 @@
 import { eventsService } from '~/server/services/eventsService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-
 	const id = event.context.params.id
-	return eventsService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		return eventsService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
