+++ server/routes/game_player_stats/[id].delete.ts
@@ -1,7 +1,10 @@
 import { statsService } from '~/server/services/statsService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return statsService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return statsService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
