+++ server/routes/leaders/[id].delete.ts
@@ -1,7 +1,10 @@
 import { leadersService } from '~/server/services/leadersService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return leadersService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return leadersService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
