+++ server/routes/news/[id].delete.ts
@@ -1,7 +1,10 @@
 import { newsService } from '~/server/services/newsService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return newsService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return newsService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
