+++ server/routes/newsletters/[id].delete.ts
@@ -1,7 +1,10 @@
 import { newslettersService } from '~/server/services/newslettersService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return newslettersService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return newslettersService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
