+++ server/routes/opponents/[id].delete.ts
@@ -1,7 +1,10 @@
 import { opponentsService } from '~/server/services/opponentsService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return opponentsService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return opponentsService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
