+++ server/routes/sms/[id].delete.ts
@@ -1,7 +1,10 @@
 import { smsService } from '~/server/services/smsService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return smsService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return smsService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
