+++ server/routes/sponsors/[id].delete.ts
@@ -1,7 +1,10 @@
 import { sponsorsService } from '~/server/services/sponsorsService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return sponsorsService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return sponsorsService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
