+++ server/routes/supporting_accounts/[id].delete.ts
@@ -1,7 +1,10 @@
 import { supportingaccountsService } from '~/server/services/supportingaccountsService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return supportingaccountsService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return supportingaccountsService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
