+++ server/routes/users/[id].delete.ts
@@ -1,7 +1,10 @@
 import { usersService } from '~/server/services/usersService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return usersService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return usersService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
