+++ server/routes/videos/[id].delete.ts
@@ -1,7 +1,10 @@
 import { videosService } from '~/server/services/videosService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return videosService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return videosService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
