+++ server/routes/votes/[id].delete.ts
@@ -1,7 +1,10 @@
 import { votesService } from '~/server/services/votesService'
 
 export default defineEventHandler(async (event) => {
-	protectEndpoint(event)
-	const id = event.context.params.id
-	return votesService.deleteOne(id)
+	if (okProtectedEndpoint(event)) {
+		const id = event.context.params.id
+		return votesService.deleteOne(id)
+	} else {
+		return 'restricted'
+	}
 })
